OC-7483:  Fixed issue with finding existing item data entry

+++ web/src/main/java/org/akaza/openclinica/controller/openrosa/processor/ItemProcessor.java
@@ -247,7 +247,7 @@ public class ItemProcessor implements Processor, Ordered {
         for (ItemData itemData: itemDataList) {
             if (itemData.getItem().getItemId() == itemId && 
                     itemData.getEventCrf().getEventCrfId() == eventCrfId && 
-                    itemData.getOrdinal() == itemOrdinal) return itemData;
+                    itemData.getOrdinal() == itemOrdinal.intValue()) return itemData;
         }
         return null;
     }
